DLQ-service/requirements.txt
fastapi
uvicorn[standard]
google-cloud-tasks